% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{icubedsim.simulation}
\alias{icubedsim.simulation}
\title{Simulate ICU occupancy}
\usage{
icubedsim.simulation(
  seed = FALSE,
  typedata = "synthetic",
  data = NULL,
  month = 2,
  nBedsStart = 40,
  icutype = "fix",
  nsimulation = 2,
  election = NA,
  cAR = NA,
  coefHE = NA,
  coefHU = NA,
  coefDE = NA,
  coefDU = NA,
  coefME = NA,
  coefMU = NA,
  coef = 1,
  error = FALSE,
  error.perc = NA,
  losprediction = "stochastic",
  CO = 0.5,
  type = "analysis"
)
}
\arguments{
\item{seed}{The seed number. FALSE if you don't want any seed parameter. A number if you want select a seed to reproduce the same outcome.}

\item{typedata}{Could be "synthetic" if you want a simulation of original data (admissionid, lengthofstay and urgency) or NA if ypu want to use your original data.}

\item{data}{It must be a data.frame contain an admissionid column, a lengthofstay column and a urgency column.}

\item{month}{They represent months over simulation is performed. Month are considered of 28 days.}

\item{nBedsStart}{It's the number of bed of the ICU. For variable ICU bed simulation this number varies over -0.125 and +0.125 of the selected one.}

\item{icutype}{It could be a "fix"bed ICU or a "variable"bed ICU.}

\item{nsimulation}{It's the number of simulation repeated.}

\item{election}{It's the relative frequency of election admission (value between 0 and 1). 1-election represents urgency admission.}

\item{cAR}{Cumulative admission ratio it's the hourly admission rate used in simulation. If you left NA it will be 0.20. cAR used in simulation it's 1.25 times of cAR selected considering ICU rejected patients.}

\item{coefHE}{It's a vector of four number which represent difference in admission rate thought daily solt time (6-12, 12-18, 18-24, 24-6) for elective patients. If yu left NA it would be c(1, 1, 1, 1). Sum must be 4.}

\item{coefHU}{It's a vector of four number which represent difference in admission rate thought daily solt time (6-12, 12-18, 18-24, 24-6) for urgent patients. If yu left NA it would be c(1, 1, 1, 1). Sum must be 4.}

\item{coefDE}{It's a vector of seven number which represent difference in admission rate thought days of the week for elective patients. If yu left NA it would be c(1, 1, 1, 1, 1, 1, 1). Sum must be 7.}

\item{coefDU}{It's a vector of seven number which represent difference in admission rate thought days of the week for urgent patients. If yu left NA it would be c(1, 1, 1, 1, 1, 1, 1). Sum must be 7.}

\item{coefME}{It's a vector of twelve number which represent difference in admission rate thought months of the year for elective patients. If yu left NA it would be c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1). Sum must be 12.}

\item{coefMU}{It's a vector of twelve number which represent difference in admission rate thought months of the year for urgent patients. If yu left NA it would be c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1). Sum must be 12.}

\item{coef}{It's a coefficient which permit to multiply all admission rate for a value simulating for example a disaster or an epidemic virus.}

\item{error}{Default=ult it's FALSE. If you decide TRUE you have to select a error.perc}

\item{error.perc}{It's the percentage value of the AR to simulate the admission rate from a normal distribution with mean the AR and sd the error.perc x AR.}

\item{losprediction}{"stochastic". Predictive algorithm it's not available in this version.}

\item{CO}{Represent the cut-off for variable model algorithm porbability to oper or close bed. Values could be between 0 and 1.}

\item{type}{Represent the type of return. It could be "analysis" for the analytic result or "result" for the origina return of the simulation}
}
\value{
rejectedAdmission

hoursRejectedAdmission

timeToAdmissionRejected

admittedAdmission

bedMatrixList

prob

plot1

hourbedfreeday

hourbedfreedaysd

meanrejected

sdrejected

meanHourRejected

sdHourRejected
}
\description{
This function simulate ICU occupancy based on admission rate coefficients, number of beds ant fix or variable ICU bed approach.
}
\examples{
icubedsim.simulation(seed = 12345, month = 8, nBedsStart = 20, icutype = "fix", coefHE = c(1, 1.5, 1, 0.5), type = "analysis")

}
